Add files via upload
This program scrapes data from Riot Games API for the inputted summoner names. 
The model then learns on the set of champions resulting in a win and set of champions that result in a loss.
The goal is for the model to predict the outcome of a game, based on the two sets of five champions on each team, by learning on the results of the players' matches.